added dropdown for projects in task form

+++ CentTask1/CentTask1/DTO/ProjectDto/GetProjectDto.cs
@@ -9,7 +9,7 @@ namespace CentTask1.DTO.ProjectDto
         public string? Description { get; set; }
         public DateTime StartDate { get; set; }
         public DateTime EndDate { get; set; }
-        public int? Budget { get; set; }
+        public double? Budget { get; set; }
         public string? ClientName { get; set; }
         public string? Status { get; set; } = string.Empty;
         public string? Manager { get; set; }
